backend/requirements.txt
# Core FastAPI and async server
fastapi>=0.104.0,<0.105.0
uvicorn[standard]>=0.24.0,<0.25.0
websockets>=12.0,<13.0

# Environment and config
python-dotenv>=1.0.0,<2.0.0

# Authentication and security
bcrypt>=4.0.1,<5.0.0
pyjwt>=2.8.0,<3.0.0

# Data validation and serialization
pydantic>=2.5.0,<3.0.0
pydantic[email]>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# Database ORM and drivers
sqlalchemy>=2.0.23,<2.1.0
psycopg2-binary>=2.9.9,<3.0.0

# Optional: For database migrations
alembic>=1.12.1,<2.0.0
faiss-cpu>=1.7.4, <2.0.0
redis>=5.0.1,<6.0.0

sentence-transformers>=2.2.2
google-generativeai>=0.3.0
numpy>=1.24.3